@@ -6,7 +6,7 @@ version: 1.2+5
 publish_to: none
 
 environment:
-  sdk: ">=2.8.1 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -19,26 +19,24 @@ dependencies:
   clima_data:
     path: ../clima_data
 
-  auto_size_text: 2.1.0
-  connectivity: 3.0.3
-  dartz: 0.10.0-nullsafety.1
-  equatable: 2.0.0
-  flutter_hooks: 0.16.0
+  auto_size_text: 3.0.0-nullsafety.0
+  dartz: 0.10.0-nullsafety.2
+  equatable: 2.0.3
+  flutter_hooks: 0.17.0
   flutter_riverpod: 0.14.0+3
   flutter_spinkit: 5.0.0
-  weather_icons: 2.0.2
-  hooks_riverpod: 0.14.0+3
-  http: 0.13.2
+  weather_icons: 3.0.0
+  hooks_riverpod: 0.14.0+4
   material_floating_search_bar:
     git:
       url: https://github.com/CentaurusApps/material_floating_search_bar.git
       ref: allow-passing-system-overlay-style
-  package_info: 2.0.0
+  package_info: 2.0.2
   riverpod: 0.14.0+3
-  shared_preferences: 2.0.5
-  font_awesome_flutter: 9.0.0
-  device_preview: 0.7.3
-  url_launcher: 6.0.6
+  shared_preferences: 2.0.6
+  font_awesome_flutter: 9.1.0
+  device_preview: 0.7.4
+  url_launcher: 6.0.9
   sizer: 2.0.13
 
 dev_dependencies:
