@@ -28,7 +28,7 @@ class SettingScreen extends HookWidget {
             }),
         title: Text(
           'Settings',
-          style: Theme.of(context).appBarTheme.textTheme.subtitle1,
+          style: Theme.of(context).appBarTheme.textTheme!.subtitle1,
         ),
       ),
       body: SingleChildScrollView(
